@@ -5,7 +5,7 @@ author: Filip Grzywok <f.grzywok@gmail.com>
 homepage: https://github.com/feelfreelinux/midi
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dev_dependencies:
   test: ^1.6.4
