@@ -1,6 +1,6 @@
 class DataChunk {
   final String id;
-  final int length;
+  final int? length;
   final List<int> bytes;
-  DataChunk({this.id, this.length, this.bytes});
+  DataChunk({required this.id, this.length, required this.bytes});
 }
