@@ -1,16 +1,16 @@
 class MidiHeader {
-  final int framesPerSecond;
-  final int ticksPerBeat;
-  final int ticksPerFrame;
+  final int? framesPerSecond;
+  final int? ticksPerBeat;
+  final int? ticksPerFrame;
   final int numTracks;
   final int format;
-  final int timeDivision;
+  final int? timeDivision;
 
   MidiHeader(
       {this.framesPerSecond,
       this.ticksPerBeat,
-      this.format,
-      this.numTracks,
+      required this.format,
+      required this.numTracks,
       this.ticksPerFrame,
       this.timeDivision});
 }
